pass tests on m1

+++ src/models.py
@@ -132,5 +132,5 @@ class Demucs(BaseModel):
 
     def predict_path(self, audio: str) -> dict:
         audio, sampling_rate = read(audio)
-        audio = torch.tensor(audio)
+        audio = torch.tensor(audio, dtype=torch.float32)
         return self.predict(audio, sampling_rate)
